Fix git url parsing issues
