+++ exporters/requirements.txt
@@ -5,3 +5,4 @@ jsonpath_ng
 jira
 pytz
 python-gitlab >= 2.4.0
+git-url-parse
